+++ tests/unittests/test_client.py
@@ -3,7 +3,7 @@ import tap_linkedin_ads.client as _client
 import tap_linkedin_ads
 import unittest
 import requests
-from datetime import datetime
+from datetime import datetime, timedelta
 import calendar
 
 @mock.patch("tap_linkedin_ads.client.LinkedinClient.write_access_token_to_config")
@@ -19,7 +19,8 @@ class TestLinkedInClient(unittest.TestCase):
         future_time = int(datetime.utcnow().timestamp()) + 88400
         mocked_response = mock.Mock()
         mocked_response.json.return_value = {
-            "expires_at": future_time
+            "expires_at": future_time,
+            "created_at": 1716560216
         }
         
         mocked_response.status_code = 200
@@ -45,6 +46,7 @@ class TestLinkedInClient(unittest.TestCase):
         mocked_response.json.return_value = {
             "access_token": "abcdef12345",
             "expires_at": future_time,
+            "created_at": 1716560216
         }
         mocked_post.return_value = mocked_response
 
@@ -65,7 +67,8 @@ class TestLinkedInClient(unittest.TestCase):
         mocked_response.status_code = 200
         mocked_response.json.return_value = {
             "access_token": "abcdef12345",
-            "expires_in": 5184000
+            "expires_in": 5184000,
+            "created_at": 1716560216
         }
         mocked_post.return_value = mocked_response
 
@@ -87,14 +90,16 @@ class TestLinkedInClient(unittest.TestCase):
         mocked_token_check_response = mock.Mock()
         mocked_token_check_response.json.return_value = {
             "access_token": "abcdef12345",
-            "expires_at": old_time
+            "expires_at": old_time,
+            "created_at": 1716560216
         }
         mocked_token_check_response.status_code = 200
 
         mocked_refresh_token_response = mock.Mock()
         mocked_refresh_token_response.json.return_value = {
             "access_token": "abcdef12345",
-            "expires_in": 5184000
+            "expires_in": 5184000,
+            "created_at": 1716560216
         }
         mocked_refresh_token_response.status_code = 200
         mocked_post.side_effect = [mocked_token_check_response, mocked_refresh_token_response]
@@ -113,3 +118,43 @@ class TestLinkedInClient(unittest.TestCase):
         client.fetch_and_set_access_token()
         actual = client.access_token
         self.assertEqual(expected_access_token, actual)
+
+    @mock.patch("tap_linkedin_ads.client.LOGGER")
+    def test_refresh_token_expires_within_month(self, mock_logger, mocked_post, mock_write_token):
+        '''
+        Ensure that we log a warning if the refresh token expires within a month
+        '''
+        client = _client.LinkedinClient('client_id', 'client_secret', 'refresh_token', 'access_token', 'config_path')
+
+        mocked_response = mock.Mock()
+        mocked_response.json.return_value = {
+            "expires_at": datetime.utcnow().timestamp(),
+            "created_at": (datetime.utcnow() - timedelta(days=334)).timestamp()
+        }
+        
+        mocked_response.status_code = 200
+        mocked_post.return_value = mocked_response
+
+        client.get_token_expires()
+        mock_logger.warning.assert_called_with(
+            "The refresh token is going to expire soon. Please re-authenticate your connection to generate a new token and resume extraction."
+        )
+
+    @mock.patch("tap_linkedin_ads.client.LOGGER")
+    def test_refresh_token_empty_expires_after_2_month(self, mock_logger, mocked_post, mock_write_token):
+        '''
+        Ensure that we do not log a warning if the refresh token expires after 2 months
+        '''
+        client = _client.LinkedinClient('client_id', 'client_secret', 'refresh_token', 'access_token', 'config_path')
+
+        mocked_response = mock.Mock()
+        mocked_response.json.return_value = {
+            "expires_at": datetime.utcnow().timestamp(),
+            "created_at": (datetime.utcnow() - timedelta(days=304)).timestamp()
+        }
+
+        mocked_response.status_code = 200
+        mocked_post.return_value = mocked_response
+
+        client.get_token_expires()
+        self.assertEqual(mock_logger.warning.call_count, 0)